requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Werkzeug==2.2.2
@tonejs/midi
tone
axios@1.3.6 --save
pyserial




